+++ daemon/src/bip39_ext.rs
