+++ daemon/src/taker.rs
@@ -8,6 +8,7 @@ use bdk::FeeRate;
 use bip39::Mnemonic;
 use clap::Parser;
 use clap::Subcommand;
+use daemon::bip39_ext::MnemonicExt;
 use daemon::bitmex_price_feed;
 use daemon::connection::connect;
 use daemon::db;
@@ -21,7 +22,6 @@ use daemon::projection;
 use daemon::seed::Seed;
 use daemon::supervisor;
 use daemon::wallet;
-use daemon::wallet_seed::MnemonicExt;
 use daemon::TakerActorSystem;
 use daemon::Tasks;
 use daemon::HEARTBEAT_INTERVAL;
